refactor: move cli actions to the command class
