+++ src/pdm/cli/actions.py
@@ -8,15 +8,10 @@ import os
 import sys
 import textwrap
 import warnings
-from argparse import Namespace
-from collections import defaultdict
-from itertools import chain
 from typing import Collection, Iterable, cast
 
-import tomlkit
 from resolvelib.reporters import BaseReporter
 from resolvelib.resolvers import ResolutionImpossible, ResolutionTooDeep, Resolver
-from tomlkit.items import Array
 
 from pdm import termui
 from pdm.cli.filters import GroupSelection
@@ -28,20 +23,14 @@ from pdm.cli.utils import (
     format_lockfile,
     format_resolution_impossible,
     get_pep582_path,
-    merge_dictionary,
-    save_version_specifiers,
     set_env_in_reg,
 )
-from pdm.environments import BareEnvironment, PythonLocalEnvironment
-from pdm.exceptions import NoPythonVersion, PdmUsageError, ProjectError
-from pdm.models.backends import DEFAULT_BACKEND
+from pdm.environments import BareEnvironment
+from pdm.exceptions import PdmUsageError, ProjectError
 from pdm.models.candidates import Candidate
-from pdm.models.python import PythonInfo
-from pdm.models.requirements import Requirement, parse_requirement, strip_extras
-from pdm.models.specifiers import get_specifier
+from pdm.models.requirements import Requirement, parse_requirement
 from pdm.project import Project
 from pdm.resolver import resolve
-from pdm.utils import cd, normalize_name
 
 
 def do_lock(
@@ -215,415 +204,10 @@ def do_sync(
         hooks.try_emit("post_install", candidates=candidates, dry_run=dry_run)
 
 
-def do_add(
-    project: Project,
-    *,
-    selection: GroupSelection,
-    sync: bool = True,
-    save: str = "compatible",
-    strategy: str = "reuse",
-    editables: Collection[str] = (),
-    packages: Collection[str] = (),
-    unconstrained: bool = False,
-    no_editable: bool = False,
-    no_self: bool = False,
-    dry_run: bool = False,
-    prerelease: bool = False,
-    fail_fast: bool = False,
-    hooks: HookManager | None = None,
-) -> None:
-    """Add packages and install"""
-    hooks = hooks or HookManager(project)
-    check_project_file(project)
-    if editables and no_editable:
-        raise PdmUsageError("Cannot use --no-editable with editable packages given.")
-    group = selection.one()
-    tracked_names: set[str] = set()
-    requirements: dict[str, Requirement] = {}
-    lock_groups = [] if project.lockfile.empty() else project.lockfile.groups
-    if lock_groups is not None and group not in lock_groups:
-        project.core.ui.echo(f"Adding group [success]{group}[/] to lockfile", err=True, style="info")
-        lock_groups.append(group)
-    if group == "default" or not selection.dev and group not in project.pyproject.settings.get("dev-dependencies", {}):
-        if editables:
-            raise PdmUsageError("Cannot add editables to the default or optional dependency group")
-    for r in [parse_requirement(line, True) for line in editables] + [parse_requirement(line) for line in packages]:
-        if project.name and normalize_name(project.name) == r.key and not r.extras:
-            project.core.ui.echo(
-                f"Package [req]{project.name}[/] is the project itself.",
-                err=True,
-                style="warning",
-            )
-            continue
-        if r.is_file_or_url:
-            r.relocate(project.backend)  # type: ignore[attr-defined]
-        key = r.identify()
-        r.prerelease = prerelease
-        tracked_names.add(key)
-        requirements[key] = r
-    if requirements:
-        project.core.ui.echo(
-            f"Adding packages to [primary]{group}[/] "
-            f"{'dev-' if selection.dev else ''}dependencies: "
-            + ", ".join(f"[req]{r.as_line()}[/]" for r in requirements.values())
-        )
-    all_dependencies = project.all_dependencies
-    group_deps = all_dependencies.setdefault(group, {})
-    if unconstrained:
-        if not requirements:
-            raise PdmUsageError("--unconstrained requires at least one package")
-        for req in group_deps.values():
-            req.specifier = get_specifier("")
-    group_deps.update(requirements)
-    reqs = [r for g, deps in all_dependencies.items() if lock_groups is None or g in lock_groups for r in deps.values()]
-    with hooks.skipping("post_lock"):
-        resolved = do_lock(
-            project,
-            strategy,
-            tracked_names,
-            reqs,
-            dry_run=True,
-            hooks=hooks,
-            groups=lock_groups,
-        )
-
-    # Update dependency specifiers and lockfile hash.
-    deps_to_update = group_deps if unconstrained else requirements
-    save_version_specifiers({group: deps_to_update}, resolved, save)
-    if not dry_run:
-        project.add_dependencies(deps_to_update, group, selection.dev or False)
-        project.write_lockfile(project.lockfile._data, False, groups=lock_groups)
-        hooks.try_emit("post_lock", resolution=resolved, dry_run=dry_run)
-    _populate_requirement_names(group_deps)
-    if sync:
-        do_sync(
-            project,
-            selection=GroupSelection(project, groups=[group], default=False),
-            no_editable=no_editable and tracked_names,
-            no_self=no_self,
-            requirements=list(group_deps.values()),
-            dry_run=dry_run,
-            fail_fast=fail_fast,
-            hooks=hooks,
-        )
-
-
-def _populate_requirement_names(req_mapping: dict[str, Requirement]) -> None:
-    # Update the requirement key if the name changed.
-    for key, req in list(req_mapping.items()):
-        if key and key.startswith(":empty:"):
-            req_mapping[req.identify()] = req
-            del req_mapping[key]
-
-
-def do_update(
-    project: Project,
-    *,
-    selection: GroupSelection,
-    strategy: str = "reuse",
-    save: str = "compatible",
-    unconstrained: bool = False,
-    top: bool = False,
-    dry_run: bool = False,
-    packages: Collection[str] = (),
-    sync: bool = True,
-    no_editable: bool = False,
-    no_self: bool = False,
-    prerelease: bool = False,
-    fail_fast: bool = False,
-    hooks: HookManager | None = None,
-) -> None:
-    """Update specified packages or all packages"""
-    hooks = hooks or HookManager(project)
-    check_project_file(project)
-    if len(packages) > 0 and (top or len(selection.groups) > 1 or not selection.default):
-        raise PdmUsageError("packages argument can't be used together with multiple -G or " "--no-default or --top.")
-    all_dependencies = project.all_dependencies
-    updated_deps: dict[str, dict[str, Requirement]] = defaultdict(dict)
-    locked_groups = project.lockfile.groups
-    if not packages:
-        if prerelease:
-            raise PdmUsageError("--prerelease must be used with packages given")
-        selection.validate()
-        for group in selection:
-            updated_deps[group] = all_dependencies[group]
-    else:
-        group = selection.one()
-        if locked_groups and group not in locked_groups:
-            raise ProjectError(f"Requested group not in lockfile: {group}")
-        dependencies = all_dependencies[group]
-        for name in packages:
-            matched_name = next(
-                (k for k in dependencies if normalize_name(strip_extras(k)[0]) == normalize_name(name)),
-                None,
-            )
-            if not matched_name:
-                raise ProjectError(
-                    f"[req]{name}[/] does not exist in [primary]{group}[/] "
-                    f"{'dev-' if selection.dev else ''}dependencies."
-                )
-            dependencies[matched_name].prerelease = prerelease
-            updated_deps[group][matched_name] = dependencies[matched_name]
-        project.core.ui.echo(
-            "Updating packages: {}.".format(
-                ", ".join(f"[req]{v}[/]" for v in chain.from_iterable(updated_deps.values()))
-            )
-        )
-    if unconstrained:
-        for deps in updated_deps.values():
-            for dep in deps.values():
-                dep.specifier = get_specifier("")
-    reqs = [r for deps in all_dependencies.values() for r in deps.values()]
-    resolved = do_lock(
-        project,
-        strategy,
-        chain.from_iterable(updated_deps.values()),
-        reqs,
-        dry_run=True,
-        hooks=hooks,
-        groups=locked_groups,
-    )
-    for deps in updated_deps.values():
-        _populate_requirement_names(deps)
-    if unconstrained:
-        # Need to update version constraints
-        save_version_specifiers(updated_deps, resolved, save)
-        for group, deps in updated_deps.items():
-            project.add_dependencies(deps, group, selection.dev or False)
-    if not dry_run:
-        project.write_lockfile(project.lockfile._data, False, groups=locked_groups)
-    if sync or dry_run:
-        do_sync(
-            project,
-            selection=selection,
-            clean=False,
-            dry_run=dry_run,
-            requirements=[r for deps in updated_deps.values() for r in deps.values()],
-            tracked_names=list(chain.from_iterable(updated_deps.values())) if top else None,
-            no_editable=no_editable,
-            no_self=no_self,
-            fail_fast=fail_fast,
-            hooks=hooks,
-        )
-
-
-def do_remove(
-    project: Project,
-    selection: GroupSelection,
-    sync: bool = True,
-    packages: Collection[str] = (),
-    no_editable: bool = False,
-    no_self: bool = False,
-    dry_run: bool = False,
-    fail_fast: bool = False,
-    hooks: HookManager | None = None,
-) -> None:
-    """Remove packages from working set and pyproject.toml"""
-    hooks = hooks or HookManager(project)
-    check_project_file(project)
-    if not packages:
-        raise PdmUsageError("Must specify at least one package to remove.")
-    group = selection.one()
-    lock_groups = project.lockfile.groups
-
-    deps, _ = project.get_pyproject_dependencies(group, selection.dev or False)
-    project.core.ui.echo(
-        f"Removing packages from [primary]{group}[/] "
-        f"{'dev-' if selection.dev else ''}dependencies: " + ", ".join(f"[req]{name}[/]" for name in packages)
-    )
-    with cd(project.root):
-        for name in packages:
-            req = parse_requirement(name)
-            matched_indexes = sorted((i for i, r in enumerate(deps) if req.matches(r)), reverse=True)
-            if not matched_indexes:
-                raise ProjectError(f"[req]{name}[/] does not exist in [primary]{group}[/] dependencies.")
-            for i in matched_indexes:
-                del deps[i]
-    cast(Array, deps).multiline(True)
-
-    if not dry_run:
-        project.pyproject.write()
-    if lock_groups and group not in lock_groups:
-        project.core.ui.echo(f"Group [success]{group}[/] isn't in lockfile, skipping lock.", style="warning", err=True)
-        return
-    do_lock(project, "reuse", dry_run=dry_run, hooks=hooks, groups=lock_groups)
-    if sync:
-        do_sync(
-            project,
-            selection=GroupSelection(project, default=False, groups=[group]),
-            clean=True,
-            no_editable=no_editable,
-            no_self=no_self,
-            dry_run=dry_run,
-            fail_fast=fail_fast,
-            hooks=hooks,
-        )
-
-
-def do_use(
-    project: Project,
-    python: str = "",
-    first: bool = False,
-    ignore_remembered: bool = False,
-    ignore_requires_python: bool = False,
-    save: bool = True,
-    venv: str | None = None,
-    hooks: HookManager | None = None,
-) -> PythonInfo:
-    """Use the specified python version and save in project config.
-    The python can be a version string or interpreter path.
-    """
-    from pdm.cli.commands.venv.utils import get_venv_with_name
-    from pdm.models.caches import JSONFileCache
-
-    def version_matcher(py_version: PythonInfo) -> bool:
-        return py_version.valid and (
-            ignore_requires_python or project.python_requires.contains(str(py_version.version), True)
-        )
-
-    hooks = hooks or HookManager(project)
-
-    selected_python: PythonInfo | None = None
-    if venv:
-        virtualenv = get_venv_with_name(project, venv)
-        selected_python = PythonInfo.from_path(virtualenv.interpreter)
-
-    if not project.cache_dir.exists():
-        project.cache_dir.mkdir(parents=True)
-    use_cache: JSONFileCache[str, str] = JSONFileCache(project.cache_dir / "use_cache.json")
-    if python:
-        python = python.strip()
-    if selected_python is None and python and not ignore_remembered:
-        if python in use_cache:
-            path = use_cache.get(python)
-            cached_python = PythonInfo.from_path(path)
-            if not cached_python.valid:
-                project.core.ui.echo(
-                    f"The last selection is corrupted. {path!r}",
-                    style="error",
-                    err=True,
-                )
-            elif version_matcher(cached_python):
-                project.core.ui.echo(
-                    "Using the last selection, add '-i' to ignore it.",
-                    style="warning",
-                    err=True,
-                )
-                selected_python = cached_python
-
-    if selected_python is None:
-        found_interpreters = list(dict.fromkeys(project.find_interpreters(python)))
-        matching_interpreters = list(filter(version_matcher, found_interpreters))
-        if not found_interpreters:
-            raise NoPythonVersion(f"No Python interpreter matching [success]{python}[/] is found.")
-        if not matching_interpreters:
-            project.core.ui.echo("Interpreters found but not matching:", err=True)
-            for py in found_interpreters:
-                info = py.identifier if py.valid else "Invalid"
-                project.core.ui.echo(f"  - {py.path} ({info})", err=True)
-            raise NoPythonVersion(
-                f"No python is found meeting the requirement [success]python {project.python_requires!s}[/]"
-            )
-        if first or len(matching_interpreters) == 1:
-            selected_python = matching_interpreters[0]
-        else:
-            project.core.ui.echo("Please enter the Python interpreter to use")
-            for i, py_version in enumerate(matching_interpreters):
-                project.core.ui.echo(f"{i}. [success]{py_version.path!s}[/] ({py_version.identifier})")
-            selection = termui.ask(
-                "Please select",
-                default="0",
-                prompt_type=int,
-                choices=[str(i) for i in range(len(matching_interpreters))],
-                show_choices=False,
-            )
-            selected_python = matching_interpreters[int(selection)]
-        if python:
-            use_cache.set(python, selected_python.path.as_posix())
-
-    if not save:
-        return selected_python
-    saved_python = project._saved_python
-    old_python = PythonInfo.from_path(saved_python) if saved_python else None
-    project.core.ui.echo(
-        f"Using Python interpreter: [success]{selected_python.path!s}[/] ({selected_python.identifier})"
-    )
-    project.python = selected_python
-    if project.environment.is_local:
-        project.core.ui.echo(
-            "Using __pypackages__ because non-venv Python is used.",
-            style="primary",
-            err=True,
-        )
-    if (
-        old_python
-        and old_python.executable != selected_python.executable
-        and isinstance(project.environment, PythonLocalEnvironment)
-    ):
-        project.core.ui.echo("Updating executable scripts...", style="primary")
-        project.environment.update_shebangs(selected_python.executable.as_posix())
-    hooks.try_emit("post_use", python=selected_python)
-    return selected_python
-
-
-def do_import(
-    project: Project,
-    filename: str,
-    format: str | None = None,
-    options: Namespace | None = None,
-) -> None:
-    """Import project metadata from given file.
-
-    :param project: the project instance
-    :param filename: the file name
-    :param format: the file format, or guess if not given.
-    :param options: other options parsed to the CLI.
-    """
-    from pdm.formats import FORMATS
-
-    if not format:
-        for key in FORMATS:
-            if FORMATS[key].check_fingerprint(project, filename):
-                break
-        else:
-            raise PdmUsageError(
-                "Can't derive the file format automatically, please specify it via '-f/--format' option."
-            )
-    else:
-        key = format
-    if options is None:
-        options = Namespace(dev=False, group=None)
-    project_data, settings = FORMATS[key].convert(project, filename, options)
-    pyproject = project.pyproject._data
-
-    if "tool" not in pyproject or "pdm" not in pyproject["tool"]:
-        pyproject.setdefault("tool", {})["pdm"] = tomlkit.table()
-    if "build" in pyproject["tool"]["pdm"] and isinstance(pyproject["tool"]["pdm"]["build"], str):
-        pyproject["tool"]["pdm"]["build"] = {
-            "setup-script": pyproject["tool"]["pdm"]["build"],
-            "run-setuptools": True,
-        }
-    if "project" not in pyproject:
-        pyproject.add("project", tomlkit.table())
-        pyproject["project"].add(tomlkit.comment("PEP 621 project metadata"))
-        pyproject["project"].add(tomlkit.comment("See https://www.python.org/dev/peps/pep-0621/"))
-
-    merge_dictionary(pyproject["project"], project_data)
-    merge_dictionary(pyproject["tool"]["pdm"], settings)
-    pyproject.setdefault("build-system", DEFAULT_BACKEND.build_system())
-
-    if "requires-python" not in pyproject["project"]:
-        python_version = f"{project.python.major}.{project.python.minor}"
-        pyproject["project"]["requires-python"] = f">={python_version}"
-        project.core.ui.echo(
-            "The project's [primary]requires-python[/] has been set to [primary]>="
-            f"{python_version}[/]. You can change it later if necessary."
-        )
-    project.pyproject.write()
-
-
 def ask_for_import(project: Project) -> None:
     """Show possible importable files and ask user to decide"""
+    from pdm.cli.commands.import_cmd import Command as ImportCommand
+
     importable_files = list(find_importable_files(project))
     if not importable_files:
         return
@@ -640,7 +224,7 @@ def ask_for_import(project: Project) -> None:
     if int(choice) == len(importable_files):
         return
     key, filepath = importable_files[int(choice)]
-    do_import(project, str(filepath), key)
+    ImportCommand.do_import(project, str(filepath), key)
 
 
 def print_pep582_command(project: Project, shell: str = "AUTO") -> None:
