+++ src/pdm/cli/commands/cache.py
@@ -16,7 +16,7 @@ class Command(BaseCommand):
     arguments = [verbose_option]
 
     def add_arguments(self, parser: argparse.ArgumentParser) -> None:
-        subparsers = parser.add_subparsers(title="Sub commands")
+        subparsers = parser.add_subparsers(title="commands", metavar="cache")
         ClearCommand.register_to(subparsers, "clear")
         RemoveCommand.register_to(subparsers, "remove")
         ListCommand.register_to(subparsers, "list")
